requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-keycloak==3.7.0
pyyaml==6.0.1
configparser==6.0.0
retry==0.9.2
urllib3==2.0.7
pydantic==2.5.0
identityutils @ git+https://github.com/eoepca/um-identity-service@v1.0.0